chore(EMS-4076): improve buyer country unit tests
