+++ src/ui/server/controllers/quote/buyer-country/index.test.ts
@@ -6,21 +6,14 @@ import getUserNameFromSession from '../../../helpers/get-user-name-from-session'
 import constructPayload from '../../../helpers/construct-payload';
 import { validation as generateValidationErrors } from '../../../shared-validation/buyer-country';
 import isChangeRoute from '../../../helpers/is-change-route';
-import getCountryByIsoCode from '../../../helpers/get-country-by-iso-code';
-import mapSubmittedEligibilityCountry from '../../../helpers/mappings/map-submitted-eligibility-country';
 import api from '../../../api';
 import mapCountries from '../../../helpers/mappings/map-countries';
-import { updateSubmittedData } from '../../../helpers/update-submitted-data/quote';
 import { mockReq, mockRes, mockCountries, mockSpyPromiseRejection } from '../../../test-mocks';
 import { Request, Response } from '../../../../types';
 
 describe('controllers/quote/buyer-country', () => {
   let req: Request;
   let res: Response;
-
-  let mockCountriesResponse = mockCountries;
-
-  const { 0: countryUnsupported, 1: countryQuoteOnline, 2: countryQuoteByEmail } = mockCountriesResponse;
 
   const mockFlash = jest.fn();
 
@@ -106,7 +99,7 @@ describe('controllers/quote/buyer-country', () => {
   });
 
   describe('get', () => {
-    let getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
+    let getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountries));
 
     beforeEach(() => {
       api.keystone.APIM.getCisCountries = getCisCountriesSpy;
@@ -124,7 +117,7 @@ describe('controllers/quote/buyer-country', () => {
       const expectedVariables = {
         ...singleInputPageVariables({ ...PAGE_VARIABLES, BACK_LINK: getBackLink(req.headers.referer), ORIGINAL_URL: req.originalUrl }),
         userName: getUserNameFromSession(req.session.user),
-        countries: mapCountries(mockCountriesResponse, req.session.submittedData.quoteEligibility[FIELD_ID]?.isoCode),
+        countries: mapCountries(mockCountries, req.session.submittedData.quoteEligibility[FIELD_ID]?.isoCode),
         submittedValues: req.session.submittedData.quoteEligibility,
         isChangeRoute: isChangeRoute(req.originalUrl),
       };
@@ -197,7 +190,7 @@ describe('controllers/quote/buyer-country', () => {
   });
 
   describe('post', () => {
-    let getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
+    let getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountries));
 
     beforeEach(() => {
       api.keystone.APIM.getCisCountries = getCisCountriesSpy;
@@ -212,159 +205,9 @@ describe('controllers/quote/buyer-country', () => {
         expect(res.render).toHaveBeenCalledWith(TEMPLATES.SHARED_PAGES.BUYER_COUNTRY, {
           ...singleInputPageVariables({ ...PAGE_VARIABLES, BACK_LINK: getBackLink(req.headers.referer), ORIGINAL_URL: req.originalUrl }),
           userName: getUserNameFromSession(req.session.user),
-          countries: mapCountries(mockCountriesResponse),
+          countries: mapCountries(mockCountries),
           validationErrors: generateValidationErrors(payload),
           isChangeRoute: isChangeRoute(req.originalUrl),
-        });
-      });
-    });
-
-    describe('when the submitted country can only get a quote via email', () => {
-      const selectedCountryisoCode = countryQuoteByEmail.isoCode;
-
-      mockCountriesResponse = [countryQuoteByEmail];
-
-      getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
-
-      beforeEach(() => {
-        req.body[FIELD_ID] = countryQuoteByEmail.isoCode;
-      });
-
-      it('should update the session with submitted data, populated with country object', async () => {
-        await post(req, res);
-
-        const selectedCountry = getCountryByIsoCode(mockCountriesResponse, selectedCountryisoCode);
-
-        const expectedPopulatedData = mapSubmittedEligibilityCountry(selectedCountry);
-
-        const expected = updateSubmittedData(expectedPopulatedData, req.session.submittedData.quoteEligibility);
-
-        expect(req.session.submittedData.quoteEligibility).toEqual(expected);
-      });
-
-      it('should add previousRoute, exitReason and exitDescription to req.flash', async () => {
-        await post(req, res);
-
-        expect(mockFlash).toHaveBeenCalledTimes(3);
-
-        expect(mockFlash.mock.calls[0]).toEqual(['previousRoute', ROUTES.QUOTE.BUYER_COUNTRY]);
-
-        const { GET_A_QUOTE_BY_EMAIL } = PAGES.QUOTE;
-        const { REASON } = GET_A_QUOTE_BY_EMAIL;
-
-        expect(mockFlash.mock.calls[1]).toEqual(['exitReason', REASON.BUYER_COUNTRY]);
-        expect(mockFlash.mock.calls[2]).toEqual(['exitDescription', REASON.BUYER_COUNTRY_DESCRIPTION]);
-      });
-
-      it(`should redirect to ${ROUTES.QUOTE.GET_A_QUOTE_BY_EMAIL}`, async () => {
-        await post(req, res);
-
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.QUOTE.GET_A_QUOTE_BY_EMAIL);
-      });
-    });
-
-    describe('when the submitted country is not found', () => {
-      beforeEach(() => {
-        req.body[FIELD_ID] = 'Country not in the mock response';
-      });
-
-      it(`should redirect to ${ROUTES.QUOTE.CANNOT_APPLY_EXIT}`, async () => {
-        await post(req, res);
-
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.QUOTE.CANNOT_APPLY_EXIT);
-      });
-    });
-
-    describe('when the submitted country is not supported', () => {
-      const selectedCountryName = countryUnsupported.name;
-      const selectedCountryIsoCode = countryUnsupported.isoCode;
-
-      beforeEach(() => {
-        req.body[FIELD_ID] = selectedCountryIsoCode;
-
-        mockCountriesResponse = [countryUnsupported];
-
-        getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
-
-        api.keystone.APIM.getCisCountries = getCisCountriesSpy;
-      });
-
-      it('should update the session with submitted data, populated with country object', async () => {
-        await post(req, res);
-
-        const selectedCountry = getCountryByIsoCode(mockCountriesResponse, selectedCountryIsoCode);
-
-        const expectedPopulatedData = mapSubmittedEligibilityCountry(selectedCountry);
-
-        const expected = updateSubmittedData(expectedPopulatedData, req.session.submittedData.quoteEligibility);
-
-        expect(req.session.submittedData.quoteEligibility).toEqual(expected);
-      });
-
-      it('should add previousRoute and exitReason to req.flash', async () => {
-        await post(req, res);
-
-        expect(req.flash).toHaveBeenCalledWith('previousRoute', ROUTES.QUOTE.BUYER_COUNTRY);
-
-        const countryName = selectedCountryName;
-
-        const { CANNOT_APPLY_EXIT } = PAGES;
-        const { REASON } = CANNOT_APPLY_EXIT;
-
-        const expectedReason = `${REASON.UNSUPPORTED_BUYER_COUNTRY_1} ${countryName}, ${REASON.UNSUPPORTED_BUYER_COUNTRY_2}`;
-
-        expect(req.flash).toHaveBeenCalledWith('exitReason', expectedReason);
-      });
-
-      it(`should redirect to ${ROUTES.QUOTE.CANNOT_APPLY_EXIT}`, async () => {
-        await post(req, res);
-
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.QUOTE.CANNOT_APPLY_EXIT);
-      });
-    });
-
-    describe('when the country is supported for an online quote and there are no validation errors', () => {
-      const selectedCountryIsoCode = countryQuoteOnline.isoCode;
-
-      const validBody = {
-        [FIELD_ID]: selectedCountryIsoCode,
-      };
-
-      beforeEach(() => {
-        req.body = validBody;
-
-        mockCountriesResponse = [countryQuoteOnline];
-
-        getCisCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
-
-        api.keystone.APIM.getCisCountries = getCisCountriesSpy;
-      });
-
-      it('should update the session with submitted data, populated with country object', async () => {
-        await post(req, res);
-
-        const selectedCountry = getCountryByIsoCode(mockCountriesResponse, selectedCountryIsoCode);
-
-        const expectedPopulatedData = mapSubmittedEligibilityCountry(selectedCountry);
-
-        const expected = updateSubmittedData(expectedPopulatedData, req.session.submittedData.quoteEligibility);
-
-        expect(req.session.submittedData.quoteEligibility).toEqual(expected);
-      });
-
-      it(`should redirect to ${ROUTES.QUOTE.BUYER_BODY}`, async () => {
-        await post(req, res);
-
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.QUOTE.BUYER_BODY);
-      });
-
-      describe("when the url's last substring is `change`", () => {
-        it(`should redirect to ${ROUTES.QUOTE.CHECK_YOUR_ANSWERS}`, async () => {
-          req.originalUrl = 'mock/change';
-
-          await post(req, res);
-
-          expect(res.redirect).toHaveBeenCalledWith(ROUTES.QUOTE.CHECK_YOUR_ANSWERS);
         });
       });
     });
@@ -373,6 +216,7 @@ describe('controllers/quote/buyer-country', () => {
       describe('when the get CIS countries API call fails', () => {
         beforeEach(() => {
           getCisCountriesSpy = mockSpyPromiseRejection;
+
           api.keystone.APIM.getCisCountries = getCisCountriesSpy;
         });
 
@@ -386,6 +230,7 @@ describe('controllers/quote/buyer-country', () => {
       describe('when the get CIS countries API call does not return a populated array', () => {
         beforeEach(() => {
           getCisCountriesSpy = jest.fn(() => Promise.resolve([]));
+
           api.keystone.APIM.getCisCountries = getCisCountriesSpy;
         });
